services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
boto3==1.34.162
